UPSTREAM: platform/chrome: cros_ec_proto: Update size arg types
cros_ec_cmd() takes 2 size arguments. Update them to be of the more
appropriate type size_t.

Suggested-by: Stephen Boyd <swboyd@chromium.org>
Link: https://lore.kernel.org/r/20220606201825.763788-4-pmalani@chromium.org
(cherry picked from commit f87e15fbf6d8cdb51f953338d41a4a52ad1aca14)

BUG=b:270507438
TEST=build

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4659340
Tested-by: Reka Norman <rekanorman@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
(cherry picked from commit 2a51937b24a7a343e444e5aa92f3e674179ce3c9)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4861732
Tested-by: ChiJen Chen <chi-jen.chen@quanta.corp-partner.google.com>
Commit-Queue: Ricky Chang <rickytlchang@chromium.org>

+++ include/linux/platform_data/cros_ec_proto.h
@@ -244,7 +244,7 @@ int cros_ec_get_sensor_count(struct cros_ec_dev *ec);
 bool cros_ec_handle_event(struct cros_ec_device *ec_dev);
 
 int cros_ec_cmd(struct cros_ec_device *ec_dev, unsigned int version, int command, void *outdata,
-		int outsize, void *indata, int insize);
+		size_t outsize, void *indata, size_t insize);
 
 /**
  * cros_ec_get_time_ns() - Return time in ns.
